Artistic-Style-Transfer/backend/requirements.txt
fastapi==0.111.0
Pillow==10.3.0
requests==2.32.3
numpy==1.24.4
torch==2.0.1
torchvision==0.15.2
uvicorn==0.30.1
